requirements.txt
Flask>=3.0.0,<4
Flask-WTF>=1.2.0,<2
python-dotenv>=1.2.1

flask-limiter